Restructure interfaces
